to enable simultaneous loading of features when viewing QTL blocks, move task to models/block
move getBlockFeaturesIntervalTask() from paths-progressive.js to block.js
Move promiseText() from paths-progressive.js to ember-devel.js
