+++ frontend/app/models/block.js
@@ -31,6 +31,9 @@ import { fcsProperties } from '../utils/data-types';
 
 import { stacks } from '../utils/stacks';
 
+import { promiseText } from '../utils/ember-devel';
+
+
 /* global d3 */
 
 /*----------------------------------------------------------------------------*/
@@ -1267,6 +1270,26 @@ export default Model.extend({
     features = pathsP.getBlockFeaturesInterval(this.id, /*all*/true);
     return features;
   }),
+
+  /** Request features for blockId, within the interval which
+   * requestBlockFeaturesInterval() derives from axisDimensions(blockId).
+   *
+   * Moved here from services/data/paths-progressive.js, to enable a request per block, in parallel.
+   */
+  getBlockFeaturesIntervalTask : task(function* (blockId, all) {
+    let
+      fnName = 'getBlockFeaturesIntervalTask',
+      features = yield this.get('pathsP').requestBlockFeaturesInterval(blockId, all);
+      if (trace_block)
+        dLog(fnName, blockId, all, promiseText(features));
+    return features;
+    /* tried .enqueue().maxConcurrency(3), but got 2 identical requests, so .drop() instead;
+     * Perhaps later: split requestBlockFeaturesInterval() into parameter gathering and request;
+     * the latter function becomes the task; store last request params with the corresponding task;
+     * check request params against last and if match then return that task perform result.
+     */
+  }).drop(),
+
 
   /** Request all features of the parent of this block which are referred to by
    * this block, via .values.flankingMarkers [].
